Markdown.pl: process "br" indicators at end of paragraph
Normally there's no point to a "<br />" tag at the end of a
paragraph as the end of the paragraph will force a break anyway.

Unless that "br" tag contains a "clear='...'" attribute.

Make sure that 3 or more spaces at the end of a paragraph actually
turns into a "<br clear='all' />" tag but at the same time make 2
spaces at the end of a paragraph just go away as it serves no
purpose.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>