Markdown API: eliminate problematic xmlcheck == 1
There was absolutely no benefit to passing in an xmlcheck value
of 1 to the Markdown/ProcessRaw API.  It was ignored and did NOT
result in any checking.

Change this so that any value other than a numeric 0 results
in XML checking when calling the API.

This makes the most sense and avoids creating obscure API bugs.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>