Markdown.pl: include yaml table lines for error msg lineno
The line number mentioned in any error message gets generated by
counting from the beginning of the non-yaml output.

Of course, the final output will include any yaml table if generated.

Adjust the line number in any error messages by the number of lines
of preceding yaml table that will be included in the output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>