Markdown.pl: tidy up some regex options
The %g_escape_table hash gets initialized in a BEGIN block and then
remains constant for the life of Markdown.pl.

Add the "o" (compile pattern once) option to all the patterns that
it appears in to perhaps squeeze out just a tiny little bit of extra
performance if possible.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>