Markdown.pl: allow '/' in auto-quoted attribute values
When auto-quoting the attribute value of something like this:

    <a href=http://example.com/test/>test</a>

Do not stop at the first "/", instead pick up the entire value.

Of course, whitespace will still terminate the value being auto-quoted.

Additionally, when checking the end of the tag to see if it's
self-closing (e.g. "<br/>") do not include the final character of
any value that may have been picked up by the auto-quoting process.

For the example above, that prevents the opening "a" tag from
mistakenly being considered self-closing just because the value
being auto-quoted happens to end in a "/" and butts right up against
the final ">" of the tag.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>