Markdown.pl: add new --us-ascii option for 7-bit output
With the new --us-ascii (aka --ascii) option enabled, any characters
with a code point value larger than 127 are output using their
equivalent numerical character entity.

This makes the output strictly US-ASCII (which is a subset of UTF-8)
and should allow it to survive almost any transport mechanism at
the expense of an increase in size that depends on how many
non-US-ASCII characters are present.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>